Added CI on gitlab
With this file, gitlab will automatically re-build the site on every
new commit. See also:
- https://bookdown.org/yihui/blogdown/gitlab-pages.html
- https://docs.gitlab.com/ee/ci/git_submodules.html

R packages are cached so they don't have to be rebuilt from source
every time. See also:
- https://stackoverflow.com/questions/53953122/gitlab-ci-cache-no-matching-files/54039355#54039355

I set the baseURL in config.toml to /CogSite/ for now so that it works
on the pre-domain gitlab page for testing. Should be changed later
when we get a domain.